fake-news-detection/requirements.txt
fastapi
uvicorn
feedparser
supabase
python-dotenv
pandas
scikit-learn
joblib
requests
pytest